refactor: restructure internal dir
