+++ internal/http/endpoints/notes.go
@@ -1,7 +1,9 @@
 package endpoints
 
-import "github.com/gin-gonic/gin"
-import httpHandlers "go-rest/internal/api/handlers"
+import (
+	"github.com/gin-gonic/gin"
+	httpHandlers "go-rest/internal/handlers"
+)
 
 func NotesRoutes(apiGroup *gin.RouterGroup, handler *httpHandlers.NotesHandler) {
 	apiGroup.GET("/get", handler.GetNotes)
