+++ internal/services/notes_service.go
