--- internal/utils/health_check.go
@@ -1,1 +1,0 @@
-package utils
