--- internal/utils/user_helper.go
@@ -1,18 +1,0 @@
-package utils
-
-import (
-	"golang.org/x/crypto/bcrypt"
-	"regexp"
-)
-
-func IsValidEmail(email string) bool {
-	// Regular expression for validating an email address
-	const emailRegex = `^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$`
-	re := regexp.MustCompile(emailRegex)
-	return re.MatchString(email)
-}
-
-func hashPassword(password string) (string, error) {
-	bytes, err := bcrypt.GenerateFromPassword([]byte(password), 14)
-	return string(bytes), err
-}
